+++ src/integration-test/java/com/synopsys/defensics/jenkins/test/RunFreestyleIT.java
@@ -24,9 +24,11 @@ import static java.lang.Thread.sleep;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.is;
+import static org.hamcrest.Matchers.notNullValue;
 import static org.hamcrest.Matchers.nullValue;
 
 import com.synopsys.defensics.apiserver.model.RunState;
+import com.synopsys.defensics.apiserver.model.RunVerdict;
 import com.synopsys.defensics.jenkins.result.HtmlReportPublisherTarget.HtmlReportAction;
 import com.synopsys.defensics.jenkins.result.ResultPackageAction;
 import com.synopsys.defensics.jenkins.test.utils.CredentialsUtil;
@@ -186,7 +188,11 @@ public class RunFreestyleIT {
 
   @Test
   public void testJobFailed() throws Exception {
-    DefensicsMockServer defensicsMockServer = new DefensicsMockServer(true, "PASS", RunState.ERROR);
+    DefensicsMockServer defensicsMockServer = new DefensicsMockServer(
+        true,
+        RunVerdict.WARNING.toString(),
+        RunState.ERROR
+    );
     defensicsMockServer.initServer(this.mockServer);
     ProjectUtils.setupProject(
         jenkinsRule,
@@ -201,10 +207,9 @@ public class RunFreestyleIT {
     FreeStyleBuild run = project.scheduleBuild2(0).get();
 
     assertThat(run.getResult(), is(equalTo(Result.FAILURE)));
-    assertThat(run.getActions(HtmlReportAction.class).size(), is(equalTo(0)));
+    assertThat(run.getActions(HtmlReportAction.class).size(), is(equalTo(1)));
     assertThat(run.getActions(HTMLAction.class).size(), is(equalTo(0)));
-    assertThat(project.getAction(HtmlReportAction.class),
-        is(nullValue()));
+    assertThat(project.getAction(HtmlReportAction.class), is(notNullValue()));
   }
 
   @Test