+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## v2023.9.0 - 2023-09-XX
+- Download report and optionally result-package if run terminated with ERROR/FATAL state
 - Update plugin reference Jenkins LTS version to 2.375.4
 - Update HTML Publisher plugin to 1.31
 - Update Credentials plugin to 1224.vc23ca_a_9a_2cb_0